Utils/PathHelper: split `normalize_path()` method
... into a `normalize_absolute_path()` and a `normalize_relative_path()` method.

Linux paths are expected to start with a leading slash and functions like `file_exists()` will not work correctly without it.
So... leading slashes should not be removed from absolute paths, but should be removed from relative paths (to allow for easy concatenation).

Splitting the method and using the appropriate version in all the right places should fix this.

Includes updated the tests.
Includes updating function calls to the method.