YoastCS rules: enforce import `use` statements for all used classes
.. in namespaced files and don't allow the use of fully qualified inline names for non-namespaced files.

Note: this sniff does not/cannot flag missing import `use` statements for the use of _unqualified_ classes inline or in docblocks as it cannot know whether that is a class within the same namespace (no `use` statement needed) or from another namespace.
This is a known limitation of sniffs no matter what.

Related to 303

Fixes 201

Impact on Yoast packages:

| Plugin/Tool       | Errors/Warnings |
|-------------------|-----------------|
| PHPUnit Polyfills | --
| WP Test Utils     | --
| YoastCS           | --
| WHIP              | 1
| Yoast Test Helper | 1
| Duplicate Post    | 1
| Yst ACF           | --
| Yst WooCommerce   | 53
| Yst News          | 24
| Yst Local         | 59
| Yst Video         | 37
| Yst Premium       | 128
| Yst Free          | 250

Note: this rule was previously already largely "silently" enforced via clean-up sweeps.