YoastCS rules: enforce proper grammar/punctuation in param description
This rule is disabled in WordPressCS as it doesn't play nice with the convoluted WP native array annotation.

For YoastCS it has been decided to apply the rule anyway. We'll need to evaluate on a case-by-case basis how to convert from the WP array annotation to another form of documenting the array format.

Related to 303

Impact on Yoast packages:

| Plugin/Tool       | Errors/Warnings |
|-------------------|-----------------|
| PHPUnit Polyfills | --
| WP Test Utils     | --
| YoastCS           | --
| WHIP              | --
| Yoast Test Helper | --
| Duplicate Post    | --
| Yst ACF           | --
| Yst WooCommerce   | 1
| Yst News          | --
| Yst Local         | 1
| Yst Video         | --
| Yst Premium       | --
| Yst Free          | 11

Note: this rule was previously already "silently" enforced via clean-up sweeps.