YoastCS rules: enforce that docblocks should have a valid `@return` tag
This rule is disabled in WordPressCS as WP Core does not want `@return void`, but consistency and explicitness of _intend_ is important, so we're going to turn it back on.

Note: `@return` tags will still not be enforced for constructors which should always be "never" methods anyway.

Related to 303

Impact on Yoast packages:

| Plugin/Tool       | Errors/Warnings |
|-------------------|-----------------|
| PHPUnit Polyfills | 57 // = Test fixtures.
| WP Test Utils     | --
| YoastCS           | --
| WHIP              | 1
| Yoast Test Helper | 18
| Duplicate Post    | 280
| Yst ACF           | 14
| Yst WooCommerce   | 172
| Yst News          | 92
| Yst Local         | 204 (but there are also still 54 functions without a docblock altogether)
| Yst Video         | 381
| Yst Premium       | 804 + 2 invalid void
| Yst Free          | 4162 + 3 invalid void (and 4 function without a docblock altogether)