YoastCS: always flag comma spacing in function calls
The `Universal.WhiteSpace.CommaSpacing` has modular error codes to allow for preventing duplicate error messages about the same issue from different sniffs.

WordPressCS excludes the "function call" related error codes from the sniff to prevent an overlap with the `Generic.Functions.FunctionCallArgumentSpacing` sniff, but this doesn't play nice with a code base using short arrays as the spacing after commas in a short array being passed as a parameter in a function call will then also be skipped.

This change prevents this issue and ensures that the spacing around those commas will be checked correctly.